requirements.txt
# Core ML and NLP dependencies
torch>=1.9.0
transformers>=4.20.0
datasets>=2.0.0
scikit-learn>=1.0.0
numpy>=1.21.0
pandas>=1.3.0

# Web scraping and text processing
requests>=2.25.0
beautifulsoup4>=4.9.0
PyPDF2>=3.0.0

# Streamlit for web interface
streamlit>=1.20.0

# Additional NLP libraries for enhanced functionality
spacy>=3.4.0
nltk>=3.8.0
textstat>=0.7.0
wordcloud>=1.9.0
plotly>=5.0.0
seaborn>=0.11.0
matplotlib>=3.5.0
gensim>=4.2.0
psutil>=5.8.0

# AI Assistant dependencies
openai>=1.0.0
python-dotenv>=1.0.0